moved page based scripts to their own dir
