+++ network-api/networkapi/mozfest/templates/mozfest/mozfest-base.html
@@ -63,3 +63,7 @@
   {# This uses the mofo main.js for now, to make sure "things work" until we (maybe never?) create a separate bundle #}
   <script src="{% static "_js/main.compiled.js" %}"></script>
 {% endblock %}
+
+{% block extra_scripts %}
+  <script src="{% static "_js/mozfest.compiled.js" %}"></script>
+{% endblock %}
