+++ network-api/networkapi/wagtailpages/templates/wagtailpages/pages/youtube_regrets_page.html
@@ -87,3 +87,7 @@
 {% block footer_block %}
 {% include "partials/footer.html" %}
 {% endblock %}
+
+{% block extra_scripts %}
+  <script src="{% static "_js/youtube-regrets.compiled.js" %}"></script>
+{% endblock %}
