+++ source/js/foundation/pages/mozfest/index.js
@@ -1,1 +1,3 @@
 export { bindEventHandlers } from "./template-js-handler";
+
+bindMozFestEventHandlers();
