+++ webpack.config.js
@@ -24,10 +24,15 @@ let rules = [
 
 let main = {
   devtool: false,
-  entry: `./source/js/main.js`,
+  entry: {
+    main: `./source/js/main.js`,
+    mozfest: `./source/js/foundation/pages/mozfest/index.js`,
+    "youtube-regrets": `./source/js/foundation/pages/youtube-regrets/index.js`,
+    "directory-listing-filters": `./source/js/foundation/pages/directory-listing-filters.js`
+  },
   output: {
     path: frontendPath,
-    filename: `main.compiled.js`
+    filename: `[name].compiled.js`
   },
   module: {
     rules
